requirements.txt
einops
biopython
torchvision
transformers==4.26.1
OmegaConf
scikit-learn
numpy
pandas
matplotlib
